feat : now todos can be searched
My own implementation of search bar. It works on key down so there is no need for pressing any search button. Searched todos are shown and searched phrase is marked on yellow. To fasten up only specific user todos are loaded at page load so there is no big loop whitch all todos that exist. Code that prints them is refactored with jquery f string. This is test version but it works, soon implement in project