feat: searchar, todo id. fix: edition and deletion
I finally managed to fix the search bar. Now when text is found, it is highlighted in one of the background colors. The most difficult task was to disable special regex characters such as '?' or '.', I managed to achieve this by checking if any of these special characters are present in the text being typed and disabling them via the escape character. Now it works perfectly. Also now all todo id's are counted based on all todo list lengths - a new todo is the last item so it gets an id of length+1 and never changes. If a todo is deleted, the id will not change. Whole class is refactored